move config to ocs
