+++ appinfo/routes.php
@@ -28,7 +28,6 @@ return [
        ["name" => "editor#public_page", "url" => "/s/{shareToken}", "verb" => "GET"],
        ["name" => "editor#users", "url" => "/ajax/users", "verb" => "GET"],
        ["name" => "editor#mention", "url" => "/ajax/mention", "verb" => "POST"],
-       ["name" => "editor#config", "url" => "/ajax/config/{fileId}", "verb" => "GET"],
        ["name" => "editor#create", "url" => "/ajax/new", "verb" => "POST"],
        ["name" => "editor#convert", "url" => "/ajax/convert", "verb" => "POST"],
        ["name" => "editor#save", "url" => "/ajax/save", "verb" => "POST"],
@@ -45,6 +44,7 @@ return [
     ],
     "ocs" => [
         ["name" => "federation#key", "url" => "/api/v1/key", "verb" => "POST"],
-        ["name" => "federation#keylock", "url" => "/api/v1/keylock", "verb" => "POST"]
+        ["name" => "federation#keylock", "url" => "/api/v1/keylock", "verb" => "POST"],
+        ["name" => "editorapi#config", "url" => "/api/v1/config/{fileId}", "verb" => "GET"]
     ]
 ];
